cameo/requirements.txt
absl-py==0.13.0
biopython==1.79
numpy==1.19.5
pandas==1.3.4
scipy==1.7.0
dm-tree==0.1.6
torch
argparse
ruamel.yaml
